requirements.txt
langchain-core==0.3.71
langchain-ollama==0.3.5
langchain-chroma==0.2.4
langchain-community==0.3.27
chromadb==1.0.15
PyPDF2==3.0.1
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
requests==2.32.4
